+++ dfetch/__main__.py
@@ -17,6 +17,7 @@ import dfetch.commands.format_patch
 import dfetch.commands.freeze
 import dfetch.commands.import_
 import dfetch.commands.init
+import dfetch.commands.remove
 import dfetch.commands.report
 import dfetch.commands.update
 import dfetch.commands.update_patch
@@ -52,6 +53,7 @@ def create_parser() -> argparse.ArgumentParser:
     dfetch.commands.freeze.Freeze.create_menu(subparsers)
     dfetch.commands.import_.Import.create_menu(subparsers)
     dfetch.commands.init.Init.create_menu(subparsers)
+    dfetch.commands.remove.Remove.create_menu(subparsers)
     dfetch.commands.report.Report.create_menu(subparsers)
     dfetch.commands.update.Update.create_menu(subparsers)
     dfetch.commands.update_patch.UpdatePatch.create_menu(subparsers)