+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Release 0.14.0 (unreleased)
 * Use github purl, repo and version for a github release archive in SBOM (#1063)
 * Allow ``dfetch freeze`` to accept project names to freeze only specific projects (#1063)
 * Edit manifest in-place when freezing inside a git or SVN superproject, preserving comments and layout (#1063)
+* Add new ``remove`` command to remove projects from manifest and disk (#26)
 
 Release 0.13.0 (released 2026-03-30)
 ====================================